Adding external database configuration.
Previously, database configurations had to be embedded in the culvert-site.xml
configuration, leading to a bunch of pain with packing hadoop configurations.

Now, configuration of the database can be just specified as a file on the
classpath. This is easier as (1) you don't need to pack the conf, (2) the
class path can just be set to include an existing database configuration,
making it easy to ensure a correct configuration is used.