Flushing, lazy init, and configuration freeze.
This commit is part bug fixes and part optimization.

Flushing addition can be considered both - it breaks if people are over-zealous
with their puts, so we abstracted that out to be simpler and more obvious.

Lazy init now applies to indices and the db adapter. Essentially, we don't
load each form the configuration each time, but instead load them everytime
the configuration is updated. This allows us to keep track of resources being
open and closing them as necessary.

Configuration freeze is tied to the previous - essentially, once the client
conf is set, you need to reset the conf (via setConf) for changes to take
effect. This ensures we can keep track of resources and that people don't do
dumb things/the right thing is obvious. This also has the added benefit of
ensuring atomicity in changes to the client's configuration.

Flushing/closing now also applies to tables/db. This ensures proper resource
management and control.

Exceptions have been simplified into a single class.